src/test/java/org/alan/algorithm/practice/baekjoon/clazz/five/BJ1208Test.java
package org.alan.algorithm.practice.baekjoon.clazz.five;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ1208Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        5 0
                        -7 -3 -2 5 8""", "1")
        );
    }
    
    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ1208().solve(input));
    }
}
